Dataset with public domain information on Hong Kong District Councillors (elected 2019).

A dataset containing public domain information of the 452 Hong Kong District Councillors elected in 2019.

hkdc

Format

A data frame with 452 rows and 33 variables:

ConstituencyCode

Constituency code for the District Council

Constituency_ZH

Constituency name in Chinese

Constituency_EN

Constituency name in English

District_ZH

District name in Chinese

District_EN

District name in English

Region_ZH

Region name in Chinese

Region_EN

Region name in English

Party_ZH

Party name in Chinese

Party_EN

Party name in English

DC_ZH

District Councillor name in Chinese

DC_EN

District Councillor name in English

FacebookURL

URL of the DC's Facebook Page

DCPageURL

URL of the DC's official page

Address

Office address of the DC

Phone

Phone number of the DC

Fax

Fax number of the DC

Email

Email address of the DC

WebsiteURL

URL of the DC's non government page

DCProjectPageURL

URL of the DC's page listing on Hong Kong Districts Info

ElectionYear

Year when the DC is elected

ElectionDate

Date when the DC is elected

CandidateNum

Candidate number of the DC when elected

Occupation

Occupation of the DC when elected

Political_ZH

Political Party of the DC when elected in Chinese

Political_EN

Political Party of the DC when elected in English

Camp_ZH

Political leaning/camp of the DC when elected in Chinese

Camp_EN

Political leaning/camp of the DC when elected in English

Vote

Number of votes won

VotePercentage

Share of vote

Gender_ZH

Gender of the DC in Chinese

Gender_EN

Gender of the DC in English

Tag_ZH

Granualar political leaning/camp of the DC when elected in Chinese

Tag_EN

Granualar political leaning/camp of the DC when elected in English

...

Source

https://www.districtcouncils.gov.hk

https://dce2019.hk01.com/